requirements.txt
streamlit
pandas
gspread
gspread-pandas
fpdf2